split internet settings webview into presentational component + story
Apply the webview-split pattern to paratext-registration's Internet settings:
the web view stays the data-loader (load/save settings via PAPI commands, the
save-and-restart flow, useWebViewState persistence) and a new
InternetSettingsForm presentational component renders the controlled form.

- Extract InternetSettingsForm (pure, props-driven): internet-use selector,
  optional proxy settings card, server selector, success/error alerts, and the
  save-and-restart button. The server/internet-use/proxy option lists, localize
  key helpers, and the string-keys list move here and are re-exported.
- internet-settings.web-view.tsx: keep all PAPI wiring and derived disabled/
  unsaved-changes logic; render <InternetSettingsForm> with the values.
- Add internet-settings.stories.tsx: Default, ProxyOnly, Restarting, SaveError.

Verified: typecheck, extensions lint, and storybook:build all clean.
